feat(docs): refactor imports and add new components for documentation
