+++ apps/docs/src/hooks/use-device.ts
